refactor: Merge types/docs from testing/auth/api_key.pyi
- Merged type hints and docstrings from `crudclient/testing/auth/api_key.pyi` into `crudclient/testing/auth/api_key.py`.

- Deleted `crudclient/testing/auth/api_key.pyi`.
